sched: tune: make dynamic_stune_boost and perfd work together
By default perfd boosts top-app schedtune to 50 on interactions
and resets to 10 when they're done. That's fine, but a bit excessive.

With dynamic stune boost we can control that behavior, but perfd is a little
fucker and overwrites everything as it see fit so we have to "block" it somehow.

But it doesn't make sense to completely rip out perfd meddling if the user
doesn't use dynamic_stune_boost in any shape or form. So boost to
dynamic_stune_boost value if it's set, otherwise leave perfd do its thing

Best of both worlds without any fuckery involved and crazy driver re-writes
as I've seen floating around...

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>